infrastructure/requirements.txt
aws-cdk-lib>=2.130.0
constructs>=10.0.0
# all CDK modules are in aws-cdk-lib — no extra packages needed
